apis/demo_api/requirements/common.txt
flask==1.1.1
flask_restful==0.3.8
flask-sqlalchemy==2.4.1
flask-cors==3.0.8
psycopg2==2.8.4.
